Binary linear layers, training binary BERT, knowledge transfer
- implemented custom sign function module with gradient as in XNOR-Net
paper
- implemented binary linear layer using that first binarizes weights
using approach from XNOR-Net paper and the above sign module
- implemented code for training a version of BERT that uses binary
linear layers everywhere except for in the embedding layer and the lm
head
- implemented knowledge transfer. We now have a binary bert
(BERT/binary) and the original BERT (BERT/original). The original BERT
is used during the forward pass in training and encoder_hidden_states
are retrieved and passed as teacher_hidden_states to binary BERT. There
the hidden states of binary BERT are retrieved and the transfer loss is
calculated and added to the mlm loss.